requirements.txt
aws-cdk-lib>=2.0.0
constructs>=10.0.0
boto3>=1.28.0
strands-agents
strands-agents-tools
yfinance
cdk_nag
ddgs